+++ src/rj_ui/src/ros2_temp/autonomy_interface.cpp
@@ -8,9 +8,9 @@ AutonomyInterface::AutonomyInterface(Context* context, rclcpp::Executor* executo
 
     executor->add_node(node_);
     status_subs_.reserve(kNumShells);
-    for (int i = 0; i < kNumShells; i++) {
+    for (size_t i = 0; i < kNumShells; i++) {
         status_subs_.emplace_back(node_->create_subscription<rj_msgs::msg::RobotStatus>(
-            radio::topics::robot_status_topic(i), rclcpp::QoS(1),
+            radio::topics::robot_status_topic(static_cast<int>(i)), rclcpp::QoS(1),
             [this, i](rj_msgs::msg::RobotStatus::SharedPtr status) {  // NOLINT
                 ConvertRx::ros_to_status(*status, &context_->robot_status.at(i));
             }));
